ecommerce-simulator/requirements.txt
Flask==3.0.0
requests==2.31.0
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.20.0
